\name{Erf}
\alias{Erf}
\alias{Erfc}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Error functions and its complement
}
\description{
Needed for compatibilty with Mathematica
}
\usage{
Erf(z)
Erfc(z)
}
\arguments{
  \item{z}{
standard normal variable
}
}
\value{
probability that a standard normal random variable is less than z or its complementary probability
}

\seealso{
\code{\link{EZL}}
}
\examples{
Erf(1.96)
Erfc(1.96)
}

\keyword{distribution }
